+++ wxc_sdk/telephony/__init__.py
@@ -133,10 +133,6 @@ class TelephonyType(str, Enum):
 
 
 class NumberListPhoneNumber(ApiModel):
-    """
-    Phone Number
-    """
-
     #: A unique identifier for the PSTN phone number.
     phone_number: Optional[str] = None
     #: Extension for a PSTN phone number.
@@ -256,10 +252,6 @@ class CallSourceInfo(ApiModel):
 
 
 class DestinationType(str, Enum):
-    """
-    Matching destination type for the call.
-    """
-
     #: Matching destination is a person or workspace with details in the hosted_user field.
     hosted_agent = 'HOSTED_AGENT'
     #: Matching destination is a calling feature like auto-attendant or hunt group with details in the hostedFeature
@@ -748,6 +740,7 @@ class TelephonyApi(ApiChild, base='telephony/config'):
         location_id: str = None,
         phone_number: str = None,
         available: bool = None,
+        assigned: bool = None,
         order: str = None,
         owner_name: str = None,
         owner_id: str = None,
@@ -762,6 +755,7 @@ class TelephonyApi(ApiChild, base='telephony/config'):
         included_telephony_type: TelephonyType = None,
         service_number: bool = None,
         reserved_number: bool = None,
+        elin_enabled: bool = None,
         org_id: str = None,
         **params,
     ) -> Generator[NumberListPhoneNumber, None, None]:
@@ -784,9 +778,13 @@ class TelephonyApi(ApiChild, base='telephony/config'):
         :type location_id: str
         :param phone_number: Search for this phone number.
         :type phone_number: str
-        :param available: Search among the available phone numbers. This parameter cannot be used along with owner_type
-            parameter when set to true.
+        :param available: Search among the available phone numbers. This parameter cannot be used along with the
+            `numberType` parameter when set to `EXTENSION`.
         :type available: bool
+        :param assigned: Return the list of phone numbers that are assigned to an owner when set to `true`. When set to
+            `false`, returns the list of phone numbers that are unassigned. This parameter cannot be used along with
+            the `numberType` parameter when set to `EXTENSION`.
+        :type assigned: bool
         :param order: Sort the list of phone numbers based on the following:lastName,dn,extension. Default sort will
             be based on number and extension in an Ascending order
         :type order: str
@@ -799,8 +797,9 @@ class TelephonyApi(ApiChild, base='telephony/config'):
         :type owner_type: OwnerType
         :param extension: Returns the list of PSTN phone numbers with given extension.
         :type extension: str
-        :param number_type: Returns the filtered list of phone numbers that contain a given type of number. `available`
-            or `state` query parameters cannot be used when `numberType=EXTENSION`.
+        :param number_type: Returns the filtered list of phone numbers that contain a given type of number.
+            `numberType` set to `EXTENSION` cannot be used along with the `available`, `assigned`, or `state` query
+            parameters. Possible input values:
         :type number_type: NumberType
         :param phone_number_type: Returns the filtered list of PSTN phone numbers that are of given phoneNumberType.
         :type phone_number_type: NumberListPhoneNumberType
@@ -826,6 +825,10 @@ class TelephonyApi(ApiChild, base='telephony/config'):
             numbers or `false` for assigned numbers; using both returns a `400` error. This parameter also cannot be
             used along with the `assigned` filter; using both returns a `400` error.
         :type reserved_number: bool
+        :param elin_enabled: When true, returns the list of phone numbers that are used as an Emergency Location
+            Identification Number (ELIN) in an emergency call scenario. When `false`, returns the list of phone
+            numbers that are not used as an ELIN.
+        :type elin_enabled: bool
         :param org_id: List numbers for this organization.
         :type org_id: str
         :return: yields :class:`NumberListPhoneNumber` instances